+++ Rules/IPv4/Inbound/HomeGroup.ps1
@@ -1,0 +1,83 @@
+
+<#
+MIT License
+
+Project: "Windows Firewall Ruleset" serves to manage firewall on Windows systems
+Homepage: https://github.com/metablaster/WindowsFirewallRuleset
+
+Copyright (C) 2019, 2020 metablaster zebal@protonmail.ch
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+#>
+
+. $PSScriptRoot\..\..\..\Config\ProjectSettings.ps1
+
+# Check requirements for this project
+Import-Module -Name $ProjectRoot\Modules\Project.AllPlatforms.System
+Test-SystemRequirements
+
+# Includes
+. $PSScriptRoot\DirectionSetup.ps1
+. $PSScriptRoot\..\IPSetup.ps1
+Import-Module -Name $ProjectRoot\Modules\Project.AllPlatforms.Logging
+Import-Module -Name $ProjectRoot\Modules\Project.Windows.ProgramInfo @Logs
+Import-Module -Name $ProjectRoot\Modules\Project.AllPlatforms.Utility @Logs
+
+#
+# Setup local variables:
+#
+$Group = "Home Group - IPv4"
+$FirewallProfile = "Private, Domain"
+
+# Ask user if he wants to load these rules
+Update-Context "IPv$IPVersion" $Direction $Group @Logs
+if (!(Approve-Execute @Logs)) { exit }
+
+# First remove all existing rules matching group
+Remove-NetFirewallRule -PolicyStore $PolicyStore -Group $Group -Direction $Direction -ErrorAction Ignore @Logs
+
+#
+# Rules for home group
+# NOTE: these rules are not used in project, needs additional tweaking and testing
+#
+
+New-NetFirewallRule -DisplayName "HomeGroup (Peer Networking Grouping)" `
+	-Platform $PlatformWin7 -PolicyStore $PolicyStore -Profile $FirewallProfile `
+	-Service p2psvc -Program $ServiceHost -Group $Group `
+	-Enabled False -Action Allow -Direction $Direction -Protocol TCP `
+	-LocalAddress Any -RemoteAddress LocalSubnet4 `
+	-LocalPort 3587 -RemotePort Any `
+	-LocalUser Any -EdgeTraversalPolicy Block `
+	-InterfaceType $Interface `
+	-Description "HomeGroup Peer Networking Grouping" `
+	@Logs | Format-Output @Logs
+
+New-NetFirewallRule -DisplayName "HomeGroup (PNRP)" `
+	-Platform $PlatformWin7 -PolicyStore $PolicyStore -Profile $FirewallProfile `
+	-Service PNRPsvc -Program $ServiceHost -Group $Group `
+	-Enabled False -Action Allow -Direction $Direction -Protocol UDP `
+	-LocalAddress Any -RemoteAddress LocalSubnet4 `
+	-LocalPort 3540 -RemotePort Any `
+	-LocalUser Any -EdgeTraversalPolicy Block `
+	-InterfaceType $Interface `
+	-LocalOnlyMapping $false -LooseSourceMapping $false `
+	-Description "HomeGroup (Pear Name Resolution Protocol)" `
+	@Logs | Format-Output @Logs
+
+Update-Logs